pyproject.toml
[project]
name = "tutorial"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aiosqlite>=0.22.1",
    "fastapi[standard]>=0.135.3",
    "pydantic-settings>=2.13.1",
    "sqlalchemy>=2.0.49",
]


[build-system]
requires = ["uv_build>=0.11.6,<0.12.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "ruff>=0.15.10",
]

[tool.fastapi]
entrypoint = "tutorial.main:app"

